pyproject.toml
[project]
name = "pss-fleet-data-api"
version = "0.0.1"
license = "MIT"
description = "An API server for Pixelstarships Fleet Data"
authors = [
    { name = "The worst.", email = "theworstpss@gmail.com" }
]
dependencies = [
    "fastapi>=0.111.0",
    "sqlmodel>=0.0.19",
    "fastapi-limiter>=0.1.6",
    "python-dateutil>=2.9.0.post0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.2",
]
readme = "README.md"
requires-python = ">= 3.12.3"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
virtual = true
dev-dependencies = [
    "autoflake>=2.3.1",
    "autopep8>=2.2.0",
    "flake8>=7.0.0",
    "flake8-bugbear>=24.4.26",
    "flake8-black>=0.3.6",
    "black>=24.4.2",
    "isort>=5.13.2",
    "pytest>=7.2.1",
    "pytest-asyncio>=0.23.7",
    "pytest-cov>=4.1.0",
    "pytest-env>=1.1.3",
    "vulture>=2.11",
    "coverage>=7.5.4",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/pss_fleet_data_api"]

[tool.autoflake]
in-place = true
remove-all-unused-imports = true
ignore-init-module-imports = true
recursive = true

[tool.black]
line-length = 150
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
src_paths = ["src"]
profile = "black"
line_length = 150
lines_after_imports = 2

[tool.vulture]

min_confidence = 100
paths = ["./src/api"]
sort_by_size = true
